% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{catllm-aliases}
\alias{catllm-aliases}
\alias{classify_survey}
\alias{extract_survey}
\alias{explore_survey}
\alias{classify_social}
\alias{extract_social}
\alias{explore_social}
\alias{classify_academic}
\alias{extract_academic}
\alias{explore_academic}
\alias{summarize_academic}
\alias{cerad_drawn_score}
\alias{classify_political}
\alias{extract_political}
\alias{explore_political}
\alias{summarize_political}
\alias{classify_web}
\alias{extract_web}
\alias{explore_web}
\alias{summarize_web}
\title{Domain-suffixed aliases for the CatLLM ecosystem}
\usage{
classify_survey(...)

extract_survey(...)

explore_survey(...)

classify_social(...)

extract_social(...)

explore_social(...)

classify_academic(...)

extract_academic(...)

explore_academic(...)

summarize_academic(...)

cerad_drawn_score(...)

classify_political(...)

extract_political(...)

explore_political(...)

summarize_political(...)

classify_web(...)

extract_web(...)

explore_web(...)

summarize_web(...)
}
\arguments{
\item{...}{Additional arguments passed to the Python function.}
}
\description{
These functions provide convenient domain-suffixed names so users can
tab-complete to find the right function. Each is a thin re-export from
the corresponding domain package.
}
